+++ MetersPanel.c
@@ -200,7 +200,7 @@ MetersPanel* MetersPanel_new(Settings* settings, const char* header, Vector* met
    this->rightNeighbor = NULL;
    this->leftNeighbor = NULL;
    Panel_setHeader(super, header);
-   for (int i = 0; i < Vector_size(meters); i++) {
+   for (size_t i = 0; i < Vector_size(meters); i++) {
       const Meter* meter = (const Meter*) Vector_get(meters, i);
       Panel_add(super, (Object*) Meter_toListItem(meter, false));
    }
